flask/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bidict==0.22.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
config==0.5.1
cycler==0.11.0
dateparser==1.1.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
fonttools==4.34.4
frozenlist==1.3.1
idna==3.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mplfinance==0.12.9b7
multidict==6.0.2
multitasking==0.0.11
netifaces==0.10.6
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
plotly==5.13.1
protobuf==4.21.3
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyqtgraph==0.12.4
python-binance==1.0.16
python-dateutil==2.8.2
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywatchman==1.4.1
regex==2022.3.2
requests==2.28.1
scipy==1.9.0
six==1.16.0
socketio==0.2.1
tenacity==8.2.2
# tvdatafeed @ git+https://github.com/StreamAlpha/tvdatafeed.git@a7034f04509b67224618917c29272796e4fff858
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.11
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.2.2
yarl==1.8.1
yfinance==0.1.74
zipp==3.15.0
